Refactor to use typescript
