+++ .gitignore
@@ -1,2 +1,3 @@
 .DS_Store
 node_modules
+tsconfig.tsbuildinfo
