+++ package.json
@@ -1,15 +1,16 @@
 {
   "name": "@point-of-sale/webusb-receipt-printer",
   "version": "2.0.0",
-  "type": "module",
   "exports": {
-      "browser": {
-          "import": "./dist/webusb-receipt-printer.esm.js",
-          "require": "./dist/webusb-receipt-printer.umd.js",
-          "script": "./dist/webusb-receipt-printer.umd.js"
-      }
+    "browser": {
+      "import": "./dist/webusb-receipt-printer.mjs",
+      "require": "./dist/webusb-receipt-printer.umd.js",
+      "script": "./dist/webusb-receipt-printer.umd.js"
+    }
   },
-  "module": "src/main.js",
+  "module": "dist/webusb-receipt-printer.mjs",
+  "main": "dist/webusb-receipt-printer.umd.js",
+  "types": "dist/webusb-receipt-printer.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/NielsLeenheer/WebUsbReceiptPrinter.git"
@@ -30,16 +31,19 @@
   },
   "homepage": "https://github.com/NielsLeenheer/WebUsbReceiptPrinter#readme",
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^24.1.0",
-    "@rollup/plugin-node-resolve": "^15.0.2",
     "@rollup/plugin-terser": "^0.4.1",
-    "rollup": "^3.20.7"
+    "@rollup/plugin-typescript": "^12.1.2",
+    "@types/w3c-web-usb": "^1.0.10",
+    "rollup": "^4.38.0",
+    "rollup-plugin-dts": "^6.2.1",
+    "tslib": "^2.8.1"
   },
   "scripts": {
     "build": "rollup -c",
     "dev": "rollup -c -w"
   },
   "files": [
-    "dist"
+    "dist",
+    "src"
   ]
 }
